Fixed the slideshow helper div
ABOUT DAMN TIME TOO. Look at the diff, it was just a bunch of width
calculations getting fucked up, I guess. Plus I reworked how content is
displayed on the page, may come at a detriment to the mobile rendeiring
of the page, but oh well. Plus I realise that I misspelt rendering, but
I'm not fixing it because Vim inside Hyper is broken af with arrow keys.