Added a static slide overlay
Aight, so this works as a static overlay for now and it looks pretty
decent ngl. All I need to do is make sure that it shows up when I hover
over the slide, then we Gucci with the Cheeky. For now though, I'll work
on making the help at the top of the page its own component. You know,
because REACT.