This boi be working now :D
Aight, so z-index is your friend mate. So turns out that Slick was
fucking me because of the placement of each slide. The last slide was
always placed on top of all the other slides like a twat. So then React
couldn't then detect the hover of the current slide. But it should be
good now, I haven't tested with any other slides though...