Fixed URL routing.
Entering a URL like victorijnr.com/blog didn't work before. But I fixed
that. I'm also working on reworking the blog part to make a request to
GitHub to pull the README for one of my repos and convert the Markdown
into HTML, instead of having it fixed at Cutex. I'll randomise which one gets chosen.