Started recent repo spotlight.
It'd just query GitHub for the 3 most recently committed to repositories
on my account. I get their information and display it. Going to switch
this to a "tab-like" view, where one of the 3 repos is selected at any
time and only its info is displayed. Also got carried away refactoring
stuff.