Changed the text in some pages.
Also added an extra easter egg. Or two. I'm not wasting the 13 seconds
it would take me to check if it was 1 or 2. Probably 2 though. Also came
up with another easter egg. It'd be the most ambitious one yet, but it'd
be pretty damn stupid, just like Lucas, if I did it. Who's Lucas? No
one, that's just a meta easter egg. ;)