Refactored everything.
Changing up the structure for this repo. I don't need the old, pre-react version of
Mii in here. I'm dumping that in its own separate repo. Instead,
everything here is the React website. It's nice, clean and I know what
I'm doing now to an extent. PLUS, I should be able to deploy this on
Netlify now instead of GitHub Pages. :D