Sorted out some build issues.
I want to rework this website completely. I like it, but it can be
better. We're learning Adobe XD so I can come up with a really nice
design, and then I'll see what I can do. It'd be nice if we can have
some TypeScript, but I'll just work with JS. For now.