Cleaning up LinkedInfoHandles
We have the logic behind altering each handle based on which route we're
on handled internally in the component. We updated react-router for
that. Don't worry, we migrated everything for it already. I'm not 100%
set on the styling to differentiate between selected and unselected
handles though. I'll be swapping those around.