Initial Content component cleanup
The other routes don't look like utter garbage now. Well, except for
cheeky. That's all kinds of fucked up right now ngl. We've sorted out
the LinkedInfoHandles, we've gone with a classic underline on the
summary. It's good enough. Next thing we want to do is sort out the
Contact page.