Here's a bottom-oriented NavigationBar
We should keep the CSS so we can easily use it for mobile. I don't want
it for desktop though. I don't feel that users will want to be looking
at the bottom of their screen frequently on web. We'll move this to the
top-left, make it look somewhat nice, and call it a day.