Deleted the Slideshow
Can you tell we're working on the projects page now? Yep. I want to put
some gource in here too. Just a gif for some projects so it's not a full
video taking ages to load the page or anything. I'll also need to decide
whether I want projects to link to a separate page or not. We'll see.
I'm leaning towards individual pages.