Made Highlights more responsive
We should look into LESS and see if we can use that here. Anyway, we
need to sort out our images and make them all a somewhat consistent
height. We're testing an image out for bu, and it's a long boi, so it
looks weird compared to others next to it.