Completed working on Highlights
I think. I'm going to work on Bloopy's page now, which should be fun. As
long as I don't get carried away and distracted by all the shit I have.
It's a lot, fam. We'll want to add more projects though, it's just
Bloopy and bu for now, but I'm sure we can add Waffle Co, and IYCASH
too.