Added auth for Spotify
We have the search endpoint working too (thanks to some copy-paste from
llaxi). I want to clean up the containers a little bit, and you can see
the start of that here too. There's also a tiny change to the Loader
container to make it look nicer since we have FontAwesome pro. That's
about it. Bye.