Added a rudimentary player
It's not styled well, nor is it contained in a single component or
anything like that. We have that refactoring and styling left to do then
it'll be perfect. We've tested it a little so far, and it's nice. We
could add a button to link to the song in Spotify, because I don't want
to use the Web Playback SDK just yet, since it requires user auth, and
that's too invasive for my little website. I forgot that I added that
example track. We can yeet that away, I don't really need it anymore.