We're nearly done with Playground
For now. We'll absolutely work on it more if I find more fun stuff to
do. We have our most-played tracks for the year, and the week being
displayed, but I want to make that random too. It's not fun having it be
the same two time frames all the time. The only other thing we need to
do is make sure our components are responsive. Once that's done, we can
work on our project pages.