Removed unnecessary CSS and jQuery
I'm also very close to giving up on a horizontal TechnologyWrapper
component. Centering a path inside an SVG can't be done with flexbox,
vertical align, or an auto margin. I've seen stuff that use a
pixel-specific way to alter them, but that's not going to be reusable. I
think I also cleaned up a few warnings here too.