Made Highlights more reusable
They can be used easily with/without a link. I made some changes in the
Projects route too; it's using the new Separator components to make it
look a little nicer, and more consistent with other parts of the
website. I'll need to do the same with Bloopy at some point too.