Initial MiniPlayer improvements
We have videos which autoplay. It's much better than using a GIF with
worse quality. I just need to add functionality to display my loader
when the video buffers. By that I mean I need to be able to figure out
when the video buffers. We have the CSS in place to show the loader as
an overlay on the video, as well as to hide the overlay if the video
isn't buffering.