Refined the outro sections
For all the projects. I made a nice little thing where the header for
the outro section for one project is mentioned in the outro section of
the previous project. Pretty much no one will notice that I say "it
won't bite" in Bloopy's outro, and then titled the outro for WaffleCo as
"It won't bite", but I like it :D