Multiple responsiveness changes
We've deployed, and the navigation menu didn't look great. Icons were
small, and, on Firefox, it could be displaced, because Firefox, on
mobile, doesn't have a consistently positioned URL bar. It's at the
bottom, and could appear or disappear, which alters the bottom of the
webpage. It wasn't nice when there was a gap between the navigation
menu, and the actual bottom of the website :(