Initial work on the collapse icon
I "fixed" the issue I had with the highlight. I say "fixed" since we're
just setting a minimum height, and letting the width do it's own thing.
It worked for SVG highlights which had a project path, but that's not
what I wanted. Anyway, I want to move the collapse icon to be inside the
HighlightImage overlay on mobile, so there may be a few more changes.