Reworked the CollapsibleSection header
It's effectively a Highlight still, except we built one within the
component, to make sure we could add the expand icon within the overlay
of the image. That's all done so I could make it look a little nicer on
mobile. Tbf, I think it's worth it, it's nicer than what we had before.